Relocate text file
