+++ src/shared/localizedText.ts
