+++ src/ssmDocument/commands/publishDocument.ts
@@ -11,10 +11,10 @@ import { AwsContext } from '../../shared/awsContext'
 import { SsmDocumentClient } from '../../shared/clients/ssmDocumentClient'
 import { ssmJson, ssmYaml } from '../../shared/constants'
 import { ext } from '../../shared/extensionGlobals'
+import * as localizedText from '../../shared/localizedText'
 import { getLogger, Logger } from '../../shared/logger'
 import { RegionProvider } from '../../shared/regions/regionProvider'
 import { getRegionsForActiveCredentials } from '../../shared/regions/regionUtilities'
-import * as genericText from '../../shared/text/generic'
 import * as picker from '../../shared/ui/picker'
 import {
     DefaultPublishSSMDocumentWizardContext,
@@ -152,8 +152,8 @@ export async function updateDocument(
                     'Would you like to make this the default version for {0}?',
                     wizardResponse.name
                 ),
-                confirm: genericText.yes,
-                cancel: genericText.no,
+                confirm: localizedText.yes,
+                cancel: localizedText.no,
             },
             window
         )
